reworked the querying and database connection part
noticed that via wsgi, database connections seem to get dropped a lot and not restarted, leading
to quite weird unrelated SSL errors between the wsgi app and postgresql. Since this is not a high traffic
app, remade so the connection is started and closed for every call made. Slow but certain. :-/

revamped the query code because the retardation that is handling having no data to return from a legitimate
as an error forcing me to catch it.. better have different methods then, with a cludgy hack to determine if we
have any rows to return from the SELECT issueing one.

God knows why I don't like python for anything that's not a shellscript..